recycle/pyproject.toml
[tool.poetry]
name = "recycle"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.6"
Django = "^2.2"
psycopg2-binary = "^2.8"
django-phonenumber-field = "^2.3"
phonenumbers = "^8.10"
PiLLow = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
